+++ src-tauri/src/repository/mod.rs
@@ -4,5 +4,6 @@ pub mod git_repository;
 pub mod github_repository;
 pub mod link_repository;
 pub mod project_repository;
+pub mod settings_repository;
 pub mod tag_repository;
 pub mod todo_repository;